Services V2: Fix content length when length > 2 GB
Third parameter of std::accumulate determines the return type, so it has
to be explicitly cast to a uint64 to support 64 bit return.

Fixes #555

+++ mythtv/libs/libmythbase/http/mythhttpranges.cpp
@@ -281,7 +281,7 @@ MythHTTPStatus MythHTTPRanges::ParseRanges(const QString& Request, int64_t Total
     }
 
     // Sum the expected number of bytes to be sent
-    PartialSize = std::accumulate(ranges.cbegin(), ranges.cend(), 0, sumrange);
+    PartialSize = std::accumulate(ranges.cbegin(), ranges.cend(), static_cast<uint64_t>(0), sumrange);
     Ranges = ranges;
 
     if (VERBOSE_LEVEL_CHECK(VB_HTTP, LOG_INFO))
